refactor: move maps subscreens to `maps` package [skip ci]
