+++ app/src/main/java/com/aliernfrog/lactool/ui/activity/MainActivity.kt
@@ -21,6 +21,11 @@ import com.aliernfrog.lactool.state.*
 import com.aliernfrog.lactool.ui.component.BaseScaffold
 import com.aliernfrog.lactool.ui.dialog.AlphaWarningDialog
 import com.aliernfrog.lactool.ui.screen.*
+import com.aliernfrog.lactool.ui.screen.maps.MapsEditScreen
+import com.aliernfrog.lactool.ui.screen.maps.MapsMaterialsScreen
+import com.aliernfrog.lactool.ui.screen.maps.MapsMergeScreen
+import com.aliernfrog.lactool.ui.screen.maps.MapsRolesScreen
+import com.aliernfrog.lactool.ui.screen.maps.MapsScreen
 import com.aliernfrog.lactool.ui.sheet.*
 import com.aliernfrog.lactool.ui.theme.LACToolTheme
 import com.aliernfrog.lactool.ui.theme.Theme
