+++ app/src/main/java/com/aliernfrog/lactool/ui/screen/maps/MapsEditScreen.kt
@@ -1,4 +1,4 @@
-package com.aliernfrog.lactool.ui.screen
+package com.aliernfrog.lactool.ui.screen.maps
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.*
