+++ app/src/main/java/com/aliernfrog/lactool/ui/screen/maps/MapsRolesScreen.kt
@@ -1,4 +1,4 @@
-package com.aliernfrog.lactool.ui.screen
+package com.aliernfrog.lactool.ui.screen.maps
 
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
